status role account

+++ src/main/java/fr/cda/controle/service/StatusService.java
@@ -4,10 +4,11 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import fr.cda.controle.beans.Status;
 import fr.cda.controle.converter.StatusConverter;
 import fr.cda.controle.dto.StatusDTO;
+import fr.cda.controle.errors.AlreadyExistException;
+import fr.cda.controle.errors.NotFoundException;
 import fr.cda.controle.repositories.StatusRepository;
 
 @Service
@@ -19,30 +20,60 @@ public class StatusService {
 	@Autowired
 	private StatusConverter statusConverter;
 
-	public StatusDTO getStatus(String type ) {
-		return statusConverter.EntityToDTO(statusRepository.findByType(type));
+	public StatusDTO getStatus(String type )  throws NotFoundException {
+		Status status= statusRepository.findByType(type);
+		if(status== null) {
+			throw new NotFoundException();
+		}
+		return statusConverter.EntityToDTO(status);
 	}
 
-	public List<StatusDTO> getAllStatus() {	
-		return statusConverter.EntityToDTO(statusRepository.findAll());
+	public List<StatusDTO> getAllStatus() throws NotFoundException {	
+		List<Status> listStatus = statusRepository.findAll();
+		if(listStatus == null) {
+			throw new NotFoundException();
+		}
+		System.out.println(statusRepository.findAll());
+		return statusConverter.EntityToDTO(listStatus);
 	}
 
-	public Status save(StatusDTO statusDTO) {
-		Status status = statusRepository.save(statusConverter.dTOToEntity(statusDTO));
-		return status;
-	}
-
-	public void delete(int id) {
-		statusRepository.deleteById(id);
+	public StatusDTO addStatus(StatusDTO statusDTO) throws AlreadyExistException{
 		
+		String type = statusDTO.getType();
+		Status status = statusRepository.findByType(type);
+		if(status == null)
+		{
+			statusRepository.save(statusConverter.dTOToEntity(statusDTO));
+		} else {
+			throw new AlreadyExistException();
+		}
+		
+		return statusDTO;
 	}
 
-	public Status update(int id,StatusDTO statusDTO) {
+	public void delete(int id) throws NotFoundException {
+		Status status = statusRepository.findById(id).get();
+		if(status == null) {
+			throw new NotFoundException();
+		}
+		else
+		{
+			statusRepository.deleteById(id);
+		}
+	}
+
+	public StatusDTO update(int id,StatusDTO statusDTO) throws NotFoundException {
 		Status status = statusConverter.dTOToEntity(statusDTO);
 		Status statusplus = statusRepository.findById(id).get();
-		statusplus.setType(status.getType());
-		statusRepository.save(statusplus);
-		return statusplus;
+		if(statusplus == null)
+		{
+			throw new NotFoundException();
+		} else {
+			statusplus.setType(status.getType());
+			statusRepository.save(statusplus);
+			return statusConverter.EntityToDTO(statusplus);
+		}
+		
 	}
